поправила кэш и конструкторы
