+++ hw15-messenger-system/src/main/java/ru/otus/backend/db/service/DbServiceUserImpl.java
@@ -3,6 +3,8 @@ package ru.otus.backend.db.service;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import ru.otus.api.cachehw.HwCache;
+import ru.otus.api.cachehw.HwListener;
 import ru.otus.api.dao.UserDao;
 import ru.otus.api.model.User;
 import ru.otus.backend.db.sessionmanager.SessionManager;
@@ -17,8 +19,19 @@ public class DbServiceUserImpl implements DBServiceUser {
 
   private final UserDao userDao ;
 
-  public DbServiceUserImpl(UserDao userDao) {
+  private final HwCache<String, User> cache;
+
+  private final HwListener<String, User> listener = new HwListener<String, User>() {
+    @Override
+    public void notify(String key, User value, String action) {
+      logger.info("key:{}, value:{}, action: {}", key, value, action);
+    }
+  };
+
+  public DbServiceUserImpl(UserDao userDao, HwCache<String, User> cache) {
     this.userDao = userDao;
+    this.cache = cache;
+    cache.addListener(listener);
   }
 
   @Override
@@ -29,6 +42,7 @@ public class DbServiceUserImpl implements DBServiceUser {
         long userId = userDao.saveUser(user);
         sessionManager.commitSession();
         logger.info("created user: {}", userId);
+        cache.put(Long.toString(user.getId()),user);
         return userId;
       } catch (Exception e) {
         logger.error(e.getMessage(), e);
@@ -38,26 +52,34 @@ public class DbServiceUserImpl implements DBServiceUser {
     }
   }
 
-
   @Override
   public Optional<User> getUser(long id) {
-    try (SessionManager sessionManager = userDao.getSessionManager()) {
-      sessionManager.beginSession();
-      try {
-        Optional userOptional = userDao.findById(id);
-        logger.info("user: {}", userOptional.orElse(null));
-        return userOptional;
-      } catch (Exception e) {
-        logger.error(e.getMessage(), e);
-        sessionManager.rollbackSession();
+    if (cache.getCache().containsKey(Long.toString(id))){
+      return Optional.ofNullable(cache.get(Long.toString(id)));
+    }else {
+      try (SessionManager sessionManager = userDao.getSessionManager()) {
+        sessionManager.beginSession();
+        try {
+          Optional <User> userOptional = userDao.findById(id);
+          logger.info("user: {}", userOptional.orElse(null));
+          cache.put(Long.toString(id), userOptional.get());
+          return userOptional;
+        } catch (Exception e) {
+          logger.error(e.getMessage(), e);
+          sessionManager.rollbackSession();
+        }
+        return Optional.empty();
       }
-      return Optional.empty();
     }
   }
 
 
   @Override
   public Optional<User> getUser(String login) {
+    for (User user :cache.getCache().values()){
+      if (user.getLogin().equals(login))
+        return Optional.ofNullable(cache.get(Long.toString(user.getId())));
+    }
     try (SessionManager sessionManager = userDao.getSessionManager()) {
       sessionManager.beginSession();
       try {
@@ -79,6 +101,7 @@ public class DbServiceUserImpl implements DBServiceUser {
       try {
           List<User> users = userDao.findAllUser();
           logger.info("user: {}", users.toString());
+          users.forEach(user -> cache.put(Long.toString(user.getId()), user));
           return users;
       } catch (Exception e) {
           logger.error(e.getMessage(), e);
